feat: add existsById before delete

+++ src/main/java/it/montano/multipersistencebackend/order/service/OrderMongoService.java
@@ -61,6 +61,9 @@ public class OrderMongoService implements OrderService {
       })
   @Override
   public void deleteOrder(@NonNull UUID orderId) {
+    if (!repo.existsById(orderId)) {
+      throw new ResourceNotFoundException("Order not found with id " + orderId);
+    }
     repo.deleteById(orderId);
   }
